requirements.txt
fastapi
uvicorn
requests
pydantic
sentence-transformers
faiss-cpu
numpy
torch
transformers